Add decode token tests, fix circular import
