+++ web/src/features/auth/slices/authenticationSlice.ts
@@ -1,8 +1,6 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit'
 
-import axios from 'api/axios'
 import { AppThunk } from 'app/store'
-import { selectToken } from 'app/rootReducer'
 import kcInstance, { kcInitOption } from 'features/auth/keycloak'
 import jwt_decode from 'jwt-decode'
 import { logError } from 'utils/error'
@@ -79,8 +77,13 @@ export const decodeRoles = (token: string | undefined) => {
   }
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   const decodedToken: any = jwt_decode(token)
-  const roles = decodedToken.resource_access['wps-web'].roles
-  return roles
+  try {
+    const roles = decodedToken.resource_access['wps-web'].roles
+    return roles
+  } catch (e) {
+    // User has no roles
+    return []
+  }
 }
 
 export const authenticate = (): AppThunk => dispatch => {
@@ -113,16 +116,4 @@ export const authenticate = (): AppThunk => dispatch => {
         dispatch(authenticate())
       })
   }
-}
-
-export const setAxiosRequestInterceptors = (): AppThunk => (_, getState) => {
-  // Use axios interceptors to intercept any requests and add authorization headers.
-  axios.interceptors.request.use(config => {
-    const token = selectToken(getState())
-    if (token) {
-      config.headers.Authorization = `Bearer ${token}`
-    }
-
-    return config
-  })
 }
